opc/ccv3/requirements.txt
# CCv3 Hackathon Dependencies

# MongoDB Atlas
motor>=3.3.0
pymongo>=4.6.0

# HTTP client
httpx>=0.24.0

# Data handling
pyyaml>=6.0.0
pydantic>=2.0.0

# Optional: Local embeddings fallback
# sentence-transformers>=2.2.0

# Web API (for Vercel deployment)
fastapi>=0.100.0
uvicorn>=0.23.0
